ansible-exporter/requirements.txt
fastapi[standard]
python-dotenv
httpx
numpy_financial
